rcar-gen5: opengl: Conditional remove opengl
OpenGL is remove by machine config, this cause user can not re-add it

+++ meta-rcar-gen5/recipes-graphics/mesa/mesa_%.bbappend
@@ -1,9 +1,11 @@
 # These below libraries are provided by gles-user-module
-PACKAGECONFIG:remove_rcar-gen5 = "egl gles"
+PACKAGECONFIG:remove:rcar-gen5 = "${@bb.utils.contains('MACHINE_FEATURES', 'gsx', 'egl gles', '', d)}"
 
-do_install:append_rcar-gen5() {
+do_install:append:rcar-gen5() {
     # Have to remove khrplatform.h file due to conflict with gles-user-module
     # even though libegl from mesa is removed
-    rm -rf ${D}${includedir}/KHR
+    if [ "${MACHINE_FEATURES}" =~ "gsx" ]; then
+        rm -rf ${D}${includedir}/KHR
+    fi
 }
